generic operation audit repo adn service
