+++ pkg/auth/user/adapter/adapter.go
@@ -17,12 +17,8 @@
 package adapter
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
 	"github.com/devtron-labs/devtron/api/bean"
 	"github.com/devtron-labs/devtron/pkg/auth/user/repository"
-	"github.com/devtron-labs/devtron/pkg/sql"
 	"strings"
 	"time"
 )
@@ -41,20 +37,4 @@ func CreateRestrictedGroup(roleGroupName string, hasSuperAdminPermission bool) b
 		Group:                   trimmedGroup,
 		HasSuperAdminPermission: hasSuperAdminPermission,
 	}
-}
-func BuildPermissionAuditModel(entityId int32, entityType repository.EntityType,
-	operationType repository.OperationType, permissionsAuditDto *bean.PermissionsAuditDto, userIdForAuditLog int32) (*repository.PermissionsAudit, error) {
-	permissionsJson, err := json.Marshal(permissionsAuditDto)
-	if err != nil {
-		errToReturn := fmt.Sprintf("error in marshalling permissions audit dto :%s", err.Error())
-		return nil, errors.New(errToReturn)
-	}
-	return &repository.PermissionsAudit{
-		EntityId:        entityId,
-		EntityType:      entityType,
-		OperationType:   operationType,
-		PermissionsJson: string(permissionsJson),
-		AuditLog:        sql.NewDefaultAuditLog(userIdForAuditLog),
-	}, nil
-
 }
