+++ pkg/operationAudit/bean/bean.go
@@ -1,0 +1,23 @@
+package bean
+
+type OperationType string
+
+const (
+	CreateOperationType OperationType = "CREATE"
+	UpdateOperationType OperationType = "UPDATE"
+	DeleteOperationType OperationType = "DELETE"
+)
+
+type EntityType string
+
+const (
+	UserEntity      EntityType = "user"
+	RoleGroupEntity EntityType = "role-group" // this is similar to permissions group
+)
+
+type SchemaFor string
+
+const (
+	UserSchema      SchemaFor = "user/v1"
+	RoleGroupSchema SchemaFor = "role-group/v1"
+)
